Move http-client tests as unit tests
